core: break af logic to own file
This is intended to keep upstream merges streamlined
by moving 'custom' logic to a non-colliding space.
